+++ src/aabb.rs
@@ -1,4 +1,4 @@
-use crate::Collides;
+use crate::{abs, Collides, Penetration};
 
 /// Axis-Aligned Bounding Box
 #[derive(Debug, Copy, Clone, PartialEq)]
@@ -20,43 +20,19 @@ impl Aabb {
         let y = Range::from_min_max(min[1], max[1]);
         Self { x, y }
     }
-
-    /// Returns by how much `self` should be moved in order to resolve penetration with `other`
-    ///
-    /// Returns `None` if the two shape are not collided
-    #[must_use]
-    #[cfg(any(feature = "std", feature = "libm"))]
-    pub fn penetration(self, other: Self) -> Option<[f32; 2]> {
-        let x = self.x.penetration(other.x)?;
-        let y = self.y.penetration(other.y)?;
-        Some(if abs(x) < abs(y) { [x, 0.] } else { [0., y] })
-    }
-
-    /// Returns the maximum penetration of `self` against the `others` shapes.
-    ///
-    /// Returns `None` if `self` does not penetrate any of the `others` shapes.
-    #[must_use]
-    #[cfg(any(feature = "std", feature = "libm"))]
-    pub fn max_penetration(self, others: impl IntoIterator<Item = Aabb>) -> Option<[f32; 2]> {
-        let mut max_magnitude = f32::MIN;
-        let mut max = None;
-        others
-            .into_iter()
-            .filter_map(move |other| self.penetration(other))
-            .for_each(|p| {
-                let magnitude = abs(p[0]) + abs(p[1]);
-                if magnitude > max_magnitude {
-                    max_magnitude = magnitude;
-                    max = Some(p);
-                }
-            });
-        max
-    }
 }
 
 impl Collides for Aabb {
     fn collides(&self, other: &Aabb) -> bool {
         self.x.collides(&other.x) && self.y.collides(&other.y)
+    }
+}
+
+impl Penetration for Aabb {
+    fn penetration(&self, other: &Self) -> Option<[f32; 2]> {
+        let x = self.x.penetration(other.x)?;
+        let y = self.y.penetration(other.y)?;
+        Some(if abs(x) < abs(y) { [x, 0.] } else { [0., y] })
     }
 }
 
@@ -74,7 +50,6 @@ impl Range {
         }
     }
 
-    #[cfg(any(feature = "std", feature = "libm"))]
     fn penetration(self, other: Self) -> Option<f32> {
         let p1 = Some(other.min - self.max).filter(|p| *p < 0.)?;
         let p2 = Some(other.max - self.min).filter(|p| *p > 0.)?;
@@ -94,17 +69,7 @@ impl From<Range> for [f32; 2] {
     }
 }
 
-#[cfg(feature = "std")]
-fn abs(v: f32) -> f32 {
-    v.abs()
-}
-
-#[cfg(all(not(feature = "std"), feature = "libm"))]
-fn abs(v: f32) -> f32 {
-    libm::fabsf(v)
-}
-
-#[cfg(all(test, any(feature = "std", feature = "libm")))]
+#[cfg(test)]
 mod tests {
     use super::*;